Import 3DE scene to Maya
The 3DEqualizer to Maya script imports the MEL script generated by 3DE and imports it in a cleaner way by locking the locators and camera, creating a proper image plane and using the dewarped images if they are found. It also scales the display of the cameras and the locators and groups it in an organized way and sets the clipping planes of the new cameras to match the scale of the scene.